INT: Add lista of GameObjects and Listeners. Bounding box are longer edges

+++ BeatEmAll/GameObjects/GameObject.h
@@ -8,7 +8,7 @@ namespace GameComponent
 	class GameObject : public Updatable, public sf::Drawable
 	{
 	protected:
-		int _x, _y;
+		float _x, _y;
 
 
 	public:
